refactor(ui): address shared-helpers review
- getErrorMessage: an empty-message Error now falls back instead of
  returning "", restoring the pre-refactor fallback semantics that
  runAction (and every mutation behind it) relies on. Add a test.
- format-time: the toLocale* wrappers fail soft to "—" on an invalid
  date, matching formatTimestamp — they're the single funnel now.
- use-copy: move to src/hooks/ (a hook, not a pure fn) and drop the
  unused timeoutMs option, so `copy` is a stable callback.
